apps/backend/parsers/sql-builder/pyproject.toml
[project]
name = "sql-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.10"
dependencies = [
    "asyncio>=3.4.3",
    "grpcio>=1.73.1",
    "grpcio-tools>=1.73.1",
    "igraph>=0.11.9",
    "pydantic-settings>=2.10.1",
    "sqlglot>=26.31.0",
]
